requirements.txt
tabulate
google-api-python-client==2.99.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1